// web/src/api/generated/model/fileDetails.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */

export interface FileDetails {
  content_url: string;
  encrypted_file_key?: string;
  encryption_format?: string;
  id: string;
  library_id: string;
  revision: number;
  size: number;
  stored_size: number;
  version_id: string;
}