// web/src/api/generated/model/problem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { ErrorDetail } from "./errorDetail.ts";

export interface Problem {
  /** A human-readable explanation specific to this occurrence of the problem. */
  detail?: string;
  /**
   * Optional list of individual error details
   * @nullable
   */
  errors?: ErrorDetail[] | null;
  /** A URI reference that identifies the specific occurrence of the problem. */
  instance?: string;
  /** Identifier of the request. Give this value to support. */
  request_id?: string;
  /** HTTP status code */
  status?: number;
  /** A short, human-readable summary of the problem type. This value should not change between occurrences of the error. */
  title?: string;
  /** A URI reference to human-readable documentation for the error. */
  type?: string;
}